Create folder current_race_screen
