+++ lib/home.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:marklin_bluetooth/controller.dart';
-import 'package:marklin_bluetooth/current_race_screen.dart';
+import 'package:marklin_bluetooth/current_race_screen/current_race_screen.dart';
 import 'package:marklin_bluetooth/race_browser.dart';
 
 class HomeScreen extends StatefulWidget {
